requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.25
protobuf==3.17.1
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.17
stripe==2.57.0
twilio==6.58.0
urllib3==1.26.4
virtualenv==20.4.4
Werkzeug==1.0.1
